initramfs: replace klibcdirs in Makefile with FORCE
'klibcdirs' was added by commit d39a206bc35d ("kbuild: rebuild initramfs
if content of initramfs changes"). If this is just a matter of forcing
execution of the recipe line, we can replace it with FORCE.

The following code is currently useless:

   $(deps_initramfs): klibcdirs

The original intent could be a hook for the klibc integration into the
kernel tree, but klibc is a separate project, which can be built
independently. Clean it up.

Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>
Reviewed-by: Greg Thelen <gthelen@google.com>
Signed-off-by: RuRuTiaSaMa <1009087450@qq.com>